Merge the JVM first and third party mappings (#15263)
Resolving #13849 will involve adding more differentiation between the "namespaces" that symbols are declared in. But as #14960 suggests, the first and third party mappings provide features that one another need in order to provide recursive matches (not to mention the thirdparty mapping trie being more memory efficient).

So, to prepare for #13849 and #14960, this change eagerly merges the first and third party mappings under the `(Frozen)TrieNode` structure. It also cleans up type and debug-goal names.