Fix ambiguity of Scala `package object`s with `class` or `object` definitions (#15265)
As described in #13849: a `type` alias inside of a `package object` is not ambiguous with a `class` or `object` declared in another file.

This change adds a mechanism to allow for multiple symbol namespaces to be used unambiguously, without consuming rules needing awareness of the namespaces. If a symbol is provided twice in the same namespace, it is ambiguous. But if it is provided in separate namespaces then it is unambiguous.

The actual heuristic used to detect use of a `package object` is not perfect: see the comment in the `_symbol_namespace` function. But the namespacing mechanism will allow us to use a more accurate namespace in future if we want to classify all symbols in the Scala parser (i.e., differentiate `type` from `class`/`object` and put them in separate namespaces directly).

Fixes #13849.