Use absolute path to shims dir in docker when tools are econfigured.
It's important that we use the absolute path to `.shims/bin` because
the use of a relative path raises an error. The error comes from the
`LookPath` function in the `golang.org/x/sys/execabs` package, which
the `docker` CLI started using as of
https://github.com/docker/cli/commit/8d199d5bba9db46b6610bd959d815ce7197402b3.
The error condition was added purposefully by the core Go team, and
it's since been ported to `os/exec` in the Go stdlib (see
https://github.com/golang/go/issues/43724).

To dynamically inject the value of `$(pwd)` at runtime, this commit
adds an additional shim script for the `docker` CLI itself. The shim
sets `PATH=$(pwd)/.shims/bin:${PATH}` before `exec`-ing the CLI.
Using a shim is a little ugly, but as far as I can tell it's the only
way to dynamically inject the working dir into the value we set for
`PATH`. Trying to set `{"PATH": "$(pwd)/.shims/bin"}` in the `env` field
for the `DockerBinary` instance we construct doesn't work because
the engine shell-escapes those values before writing them to `__run.sh`.

[ci skip-rust]

[ci skip-build-wheels]