Improve error message for resolves x interpreter constraints issues (#15314)
Improves https://github.com/pantsbuild/pants/issues/14863 a little bit.

Note that this does _not_ fix the issue: the biggest problem is we don't say which targets are the problem. As explained at https://github.com/pantsbuild/pants/issues/14863#issuecomment-1116369327, that is going to be trickier to pull off and too hard to cherry-pick. This makes an initial first step.

[ci skip-rust]
[ci skip-build-wheels]