Index resources in BSP, and expose transitive dependencies in `compile` to allow tests to consume them (#15390)
#15347 exposed only the direct classpath entries of the BSP `BuildTarget`. This was incorrect for two reasons:
1. `resources` do not have an associated `BuildTarget` of their own (they do not have an [LSP language id](https://microsoft.github.io//language-server-protocol/specifications/lsp/3.17/specification/#textDocumentItem)), and are instead attached to a nearby `BuildTarget`.
2. The transitive dependencies of a `BuildTarget` are not exposed as `BuildTarget` level module dependencies (yet).

This change adjusts the `BSPCompileRequest` implementations for Java and Scala to expose their transitive dependencies, to address both of these issues. #15051 will remove this transitive exposure in favor of module-level dependencies for source code. But it will also need to attach `resources` to a nearby `BuildTarget` deterministically, and expose them as part of that target's `class_directory`.

This change also implements `ResourcesRequest`, although it is _not_ used to add resources to a classpath: only to index them.

Resources are now indexed, and tests can consume the transitive classpath (including resources). Fixes #15050.

[ci skip-rust]
[ci skip-build-wheels]