Support the `{chroot}` env placeholder for all interactive processes that are not executed in the workspace. (Cherry-pick of #15401) (#15405)
The fix #15340 only addressed regular `Process` requests, with the intention of #15341 to leverage the support from the `run` goal, however `publish` does not use that code path, and thus does not yet have support for the `{chroot}` placeholder for env values.

This PR introduces generic support for the `{chroot}` placeholder also for `InteractiveProcess` requests that are not to be executed in the workspace. (the `run` goal executes the interactive processes in the workspace).