Fix boolean logic when merging interpreter constraints.
Previously we eagerly failed when ANDing two constraints from
different interprerter types. Now we represent that as a special
"impossible" requirement, because the final boolean expression
may still be possible on other OR branches.

E.g., previously we failed on

(CPython==3.7.* OR PyPy>=43.0) AND (PyPy==43.5)

whereas now we correctly merge this as

PyPy>=43.0,==43.5

Where the impossible branch of (CPython==3.7.* AND PyPy==43.5)
is ignored in the final OR.

We now error only if the final OR has no possible branches.

We also now correctly handle empty constraints. They should represent
"no constraints", i.e., "any interpreter is allowed", and so shouldn't
be in the final OR.

E.g., previously we merged (() AND (>=3.7)) as () (because the cross-product
we used in the computation was empty), and now we correctly merge those as
(>=3.7).

Also adds an interpreter_constraints field to python_distribution, to make
it easy to parametrize a wheel on multiple interpreter constraints.

[ci skip-rust]

[ci skip-build-wheels]