Fix `tags` to work with filesystem arguments, `overrides` field, and `python_tests` (#15394)
Closes https://github.com/pantsbuild/pants/issues/14977 and https://github.com/pantsbuild/pants/issues/11123, along with a bug that `python_tests` was moving rather than copying the `tags` field.

As explained in https://github.com/pantsbuild/pants/issues/14977, it is necessary to filter targets after we have replaced target generators with generated targets (`UnexpandedTargets -> Targets`). The cleanest way I could think to do that is adding `FilteredTargets`.

We must still proactively filter inside `AddressSpecs` and `FilesystemSpecs`, as explained in a code comment, to ensure that project introspection goals like `list` behave properly.

Note that `--changed-since` already worked, unlike filesystem arguments, because we convert the result of it into `AddressSpecs`.

[ci skip-rust]

[ci skip-build-wheels]