Support for sharding tests.
Provides generic support for sharding target roots, but only
exposes it to users for the test goal, for now.

It may be applicable to other goals, such as package, so we
can consider adding it there in the future.

But as I'm not sure if there might be unintended consequences
or confusion when applied generally, this is not a global option.

Note that only root targets are sharded. Obviously once we start
pulling in dependencies we cannot shard in any sensible way.

[ci skip-rust]

[ci skip-build-wheels]