Only infer an __init__.py dep on a python_sources-generated target.
Or, to be precise, anything with a PythonSources field.

If the __init__.py also belongs to a files() target, say because
the repo consumes itself as test data, then previously that would
cause us to infer a production dep on the "file as testdata", as
well as any python_sources() target.

Now we check that we're inferring these deps on the intended
python_sources target.

Admittedly this is an unlikely scenario, but is one I have hit
in practice, so we might as well be robust.

[ci skip-rust]

[ci skip-build-wheels]