jvm: fix incorrect check for root-only `ClasspathEntryRequest` implementations (Cherry pick of #15494) (#15497)
The conditional logic which checked for a root-only `ClasspathEntryRequest` implementation in a non-root position looked at a loop variable which could be unrelated to the compatible `ClasspathEntryRequest` implementation at issue. 

Solution: Check the root-only status of the compatible `ClasspathEntryRequest` implementation and avoid checking the loop variable.

h/t to @stuhood for spotting the bug. https://github.com/pantsbuild/example-kotlin/pull/1#issuecomment-1127985443

[ci skip-rust]

[ci skip-build-wheels]