Add initial Sphinx support with `sphinx_project` target (#15512)
This invokes `sphinx-build` to create a site. Follow up improvements:

1. Support Markdown and Myst.
2. Support generating from `.py` files. We need to model dependencies properly for this.
3. Wire up to more of Sphinx's options, e.g. the output mode.
4. Add `tailor` support, blocked by https://github.com/pantsbuild/pants/issues/15510.

## Target modeling

I at first tried to use `sphinx_source` and `sphinx_sources`. But this didn't make sense and was hard to implement. In Sphinx, a docsite is fairly atomic -- even though you can technically only generate one or two files. For example, we _must_ have the `conf.py` file anytime we invoke `sphinx-build`.

The downsides of this current approach—an atom target that encompasses the whole site—is there is no way to only build a single file and we also don't show the dependencies between individual files with project introspection. Those both seem acceptable, and result in less boilerplate for users and a much easier implementation for us.

## `package` goal

We considered: `export-codegen`, a new `gen-docs` goal, and `package`. 

The consensus was for `package` because it:

* semantically describes the action: converting the atomic raw sources into an asset
* has nice integrations like `publish`, `runtime_package_dependencies`, and `archive`
* avoids the proliferation of goals