Add `--unmatched-cli-globs`, which before was controlled by `--owners-not-found-behavior` (#15556)
Right now, `--owners-not-found-behavior` controls whether we error or not on unmatched file arguments, like `./pants test spot_the_typoo.py`. But we need to get rid of that option (https://github.com/pantsbuild/pants/issues/15529) as it blocks target-less goals like `count-loc` from properly working with file arguments. We have a bad coupling.

The fix is to add an option `--unmatched-cli-globs`. The name mirrors `--unmatched-build-file-globs` from https://github.com/pantsbuild/pants/pull/15555.

This option is not yet fully wired up—it only works for `FileLiteralSpec` and `FileGlobSpec`, but not yet `DirGlobSpec` (`dir:`) and `RecursiveGlobSpec` (`dir::`). That will happen in a followup.

[ci skip-rust]
[ci skip-build-wheels]