scala: use test lockfiles for most scalac tests (#15559)
Use test lockfile fixtures for most of the `scalac` tests. 

Note: The `test_compile_with_multiple_scala_versions` test demonstrates the limits of the current structure of the test lockfiles support: It needs multiple lockfiles, but `jvm_lockfile` is essentially a singleton because the relevant `pytest.mark.jvm_lockfile` can only be applied once. Separate fixture functions do not help, each with their own `pytest.mark.jvm_lockfile`, because those marks are ignored for fixture functions.

So the test lockfile fixture support will need to be reworked in the future, but this PR still improves the code by removing most of the code encoding lockfiles in Python.

[ci skip-rust]

[ci skip-build-wheels]