Don't error when `dir:` and `dir::` have no targets, as long as the directories exist (#15570)
This is necessary so that `./pants count-loc ::` works in a repository without any BUILD files.

We still, however, want to error when doing `./pants count-loc does-not-exist::`. That can be turned off via `--unmatched-cli-globs`. To pull this off, we create a second set of `PathGlobs` solely for validation. Closes https://github.com/pantsbuild/pants/issues/15558.

[ci skip-rust]
[ci skip-build-wheels]