`update-build-files` now takes CLI args for which BUILD files to change (#15581)
First part of https://github.com/pantsbuild/pants/issues/15563. You can now do, for example:

* `./pants update-build-files ::`
* `./pants update-build-files dir/BUILD`

Note that this deprecates leaving off arguments. Even if we think that may be desirable, I strongly encourage us to be consistent with all other goals. That will leave the semantics of "no args" open to us to change in the future, e.g. to mean something like `--changed-since=implicit`. If we are not consistent, then that semantics change will be much harder to pull off.

[ci skip-rust]
[ci skip-build-wheels]