Have `generate_docs` error on bad pantsbuild URLs (Cherry-pick of #15592)
Fixes #11578 by erroring if we get a non-200 status on a pantsbuild.org URL. Also fixed a few URLs it caught.

I originally wrote this code in a way that we'd validate all URLs (as @Eric-Arellano suggests), however it choked on our URL templates ("https://github.com/bufbuild/buf/releases/download/{version}/buf-{platform}.tar.gz") and I didn't want to throw the baby out with the bathwater trying to fix that.

Additionally, this will consistently fail on main, due to the docs not existing yet, so added --skip-check-urls to skip the checking.

I shied away from doing this kind of validation in a test, because that means the tests would rely on network connection, which I'm not a big fan of (and if we're validating any URL, the test could get very "flaky").

Also includes a little refactoring.

[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]