Use new, consistent CLI args semantics for `tailor` if `use_deprecated_cli_args_semantics = false` (#15612)
Closes https://github.com/pantsbuild/pants/issues/15563. You can now run `./pants tailor ::` and  even `./pants --changed-since=HEAD tailor`!

Rather than adding a dedicated option to `tailor`, we piggy-back off of the global option `use_deprecated_cli_args_semantics`. This will make it easier for new users to do the correct thing right away. We'll have them set this when creating the initial BUILD file.

## Quirk: file literals

Usually, `./pants tailor dir/f.py` would imply only consider if you need to add a target for `f.py`. Even if `f_test.py` needs a target, don't consider it.

However, that is pretty hard for us to implement and I don't think it's desirable. For example, I don't think it's desirable to create `python_sources(source=["f.py"])` when we normally would have used the default source globs; that is not future-proof. It is especially egregious with `./pants --changed-since=HEAD tailor`, where you presumably want to add targets for the whole directory, not just the new file.

[ci skip-rust]
[ci skip-build-wheels]