Remove `-d` as shorthand for `--logdir` (#15621)
First off, I don't think this option actually does anything currently... https://github.com/pantsbuild/pants/issues/15004#issuecomment-1135395887

Even if it did, `-d` is really confusing. It's not an enum option, so the user sets to an open-ended value like `./pants -dmy_dir`. It's clearer to do `./pants --logdir=my_dir`. This is unlike `-linfo`/`-ldebug`, which is more intuitive.

Fixing this also helps to free up the `-` symbol so that we can use it for ignore specs as an improvement over `!`, which would require shell escaping. https://github.com/pantsbuild/pants/issues/15539

[ci skip-rust]