Don't allow multiple flag names for the same option (#15627)
We never use two flags for the same option anymore, except for `-l` and `--level`. In https://github.com/pantsbuild/pants/pull/10982, we moved away from this approach under the argument that there should be one way to refer to the option. Otherwise users get confused whether things are the same etc.

This enforces that expectation with the new Options API, but not yet the underlying `parser.py` which still needs to accommodate `-l` and `--level` being the same thing.

[ci skip-rust]
[ci skip-build-wheels]