Make the global --no-process-cleanup conserve the `run` tmpdir.
It seems intuitive that it should, and the --no-run-cleanup option is still
available for the more selective behavior of just preserving the run
process's tmpdir, rather than all process tmpdirs.

Also fixes the names of a couple of unrelated copypasta rules. This didn't justify
its own change and CI run.

[ci skip-rust]

[ci skip-build-wheels]