Remove unused `is_default_target`
The idea of "default" target is not relevant. All we care about now is: was the name explicitly set or not. But it's not safe to store that information in the `Address` because `Address` can be constructed in several places, including from CLI Specs. Instead, we can only gather this information from `TargetAdaptor` (BUILD file parsing), then store it on the `Target`

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]