Handle venv path special chars in coursier fetch
On one system (a gentoo linux  box), uname reports a very verbose
processor name which is then used to build the pants venv path. That
verbose processor name includes "Intel(R)" and other symbols including
parentheses and at: `()@`

The coursier fetch script did not escape the python_path, so running
`./pants lint ::` on the pants codebase results in this error:

```
__coursier/coursier_fetch_wrapper_script.sh: line 11: syntax error near unexpected token `R'
```

The full path that triggered this error is:
```
/home/cognifloyd/.cache/pants/pants_dev_deps/Linux.x86_64.Intel(R).Core(TM).i7-3610QM.CPU.@.2.30GHz.py37.venv/bin/python
```

This PR fixes that by using `shlex.quote()` on paths inthe
coursier script.

[ci skip-rust]
[ci skip-build-wheels]