Record when the `name=` field is explicitly set  (#15693)
This is the first step to the proposal in https://docs.google.com/document/d/1jSCJ3pwvIMg5duafsr8KmiLOO_2ARMaYCSrb8K85JQc/edit#

We need to know if `name=` was left off or not so that we can deprecate the 5 ambiguous cases from the proposal.

Note that we must store this information on `TargetAdaptor` and `Target`, _not_ on `Address`. This is because `Address`es may be constructed from several sources, including `Specs`. We must consult the original `TargetAdaptor` from the BUILD file to know if it was set or not.

Generated targets don't bother with recording this. We need to consult the original generator target, anyways, to know if it was an illegal case. Not bothering saves some boilerplate and requiring plugin authors to make changes.

Finally, this fixes some target generation rules that weren't properly passing `UnionMembership`, which meant that plugin fields were being ignored for them.

[ci skip-rust]