Support running pex_binary targets directly on the in-repo sources. (#15689)
Some processes, such as Django's makemigrations use the __file__ and/or __path__ of modules to generate new paths to write to, relative to those. If run against sources in the tmpdir, those files will be written into the tmpdir and then cleaned up.

Running "inline" is necessary in some cases, and whether it is necessary or not is a property of the binary (does it need to know the true locations of the source files it runs on), and so the sandbox-ness is a property of a pex_binary target rather than a command-line option.

The consequences of running inline are that more code will be accessible on the sys.path - basically all the code in the relevant source roots. Theoretically this could lead to different behavior than running sandboxed. But this is unlikely to be a problem in practice, especially for the types of binary we're talking about. They already implicitly expect to run with that context, so if there is a difference in behavior then arguably the sandboxed behavior is the wrong one.

Note that this concept of running "inline" is Python-specific, not generic to the run goal. There is no meaning to running inline for a compiled language.

[ci skip-rust]

[ci skip-build-wheels]