Fix `--changed-since` with directories unknown to Pants (#15714)
Fixes a regression from https://github.com/pantsbuild/pants/pull/15570. When you had a changed file in a directory Pants does not find because it was deleted or it's in `--pants-ignore`, we would error.

Note that we still error when you explicitly run on the directory, like `./pants list does-not-exist`, per `--unmatched-cli-globs`. This only impacts `--changed-since`.

[ci skip-rust]