Correctly describe the origin of invalid specs, e.g. `--paths-from` (#15730)
Before:

```
❯ ./pants paths --from=fake/f.py --to=src/python/pants/util/strutil.py
12:54:23.99 [ERROR] 1 Exception encountered:

  Exception: Unmatched glob from CLI arguments
```

After:

```
❯ ./pants paths --from=fake/f.py --to=src/python/pants/util/strutil.py
12:54:23.99 [ERROR] 1 Exception encountered:

  Exception: Unmatched glob from the option `--paths-from`: "fake/f.py"
```

It was a lie to hardcode "CLI arguments". 

Most of these rules should never error because we only error if the directory of the CLI specs do not exist, and we programmatically compute the specs based on already existing targets. But still, this gives us much better debugging if those assumptions are violated.

This also unblocks us from https://github.com/pantsbuild/pants/issues/14468. The `Specs` will be able to pass down their `description_of_origin` to `AddressInput` for much better errors with invalid `AddressLiteralSpec`s.

[ci skip-build-wheels]