`WrappedTarget` now requires `WrappedTargetRequest` (#15789)
See https://github.com/pantsbuild/pants/pull/15759 for the motivation as prework for https://github.com/pantsbuild/pants/issues/14468.

The majority of call-sites here are infallible: we are going from something like `Target -> DependenciesRequest -> WrappedTarget`. That is, we know the target already exists and it's infallible. This suggests that our APIs may need to be redesigned so we don't have to re-request the target, e.g. fix our dependency inference API with https://github.com/pantsbuild/pants/issues/15400.

[ci skip-rust]
[ci skip-build-wheels]