[internal] better align test JVM lockfile support with pytest concepts (#15804)
The JVM test lockfile support had a significant limitation where only one `jvm_lockfile` fixture can be passed into a test function. This is because of a limitation with pytest marks where a mark can only be applied to a test function and not to fixture functions. Thus, the `jvm_lockfile` fixture was essentially a singleton because it was not possible to declare separate fixtures which were marked individually with the `pytest.mark.jvm_lockfile`.

This PR rewrites the test lockfile support to operate like regular pytest fixture functions. The main caveat is that the test lockfile definition must be placed in a separate fixture function to allow extracting it easily. Also, the file containing a definition must use `from __future__ import annotations` so that the type annotations are available to the extractor.

With this PR, there is no longer any need to register `JvmLockfilePlugin` in a `conftest.py file`, which further simplifies using the test lockfile support.

Moreover, it allows using multiple lockfiles in the `test_compile_with_multiple_scala_versions` test which was not possible without this PR.

[ci skip-rust]

[ci skip-build-wheels]