Add `__defaults__` BUILD file symbol (#15836)
Provide default field values for a set of targets in a particular subtree of a project using the `__defaults__` symbol, using the same syntax as for the `overrides` field.

Example:

```python
# BUILD file
__defaults__(
  {
    pyhon_sources: dict(sources=["*.py"]),
    (python_tests, python_testutils): dict(tags=["some-tag"]),
  },
  all=dict(description="Copyright (c) 2022, Company Ltd."),
)

...
```

The `__defaults__` declaration should go at the top of the BUILD file, in order to apply to the targets in the same file. All BUILD files in subdirectories inherit the same defaults by default, but can be overridden/extended with their own `__defaults__`.

Closes #13767.
Discussion: #15809.