Skip CI checks on docs-only PRs. (#16041)
A new job detects whether the change is "docs only", and all the other jobs
run only if "docs only" is false.

The complication is that we have a branch protection rule on main that prevents
merges unless a CI job named "Merge OK" passes. Today that job depends on all
the other jobs (running tests etc.) and if we skip those jobs, the "Merge OK" job
will not run and so we won't be able to merge.

Therefore we generate two "Merge OK" jobs: one that depends on all the checks but
runs only if "docs only" is false, and another that runs with no other dependencies, but
only if "docs only" is true.

Note that we cannot implement this using GitHub Action's path filtering, because that
provides only "run a workflow if any path is under a dir" or "skip a workflow if all paths
are under a dir". It does not provide "run a workflow if all paths are under a dir",
which is what we'd need for the "Merge OK" logic above.