Allow deducing flag name from option attribute name (#16040)
Removes the need for specifying the flag name to options, as it will default to the attribute name kebab-cased (and underscore stripped).

Why? Well DRY for one. Avoid typos for two. But for three, because its nice to have the code you're reading be immediately mentally mappable to the flag.

Using `generate_docs` I can confirm there is 0 diff from `main` :muscle: 

[ci skip-rust]