Write to caches during backtracking (#16078)
As reported in #16009, processes were not being written to caches during backtrack attempts, which meant that future attempts to use the same partial/invalid cache entry would also need to backtrack.

To resolve this, we disable reading from all caches during backtrack attempts, but continue writing to them (if enabled) to allow partial/invalid entries to be repaired.

Fixes #16009.

[ci skip-build-wheels]