Upgrade Pex to 2.1.95.
This picks up several fixes that make PEX environments more robust to
launch and use. It also picks up a new `--target-system` feature for
universal resolves which is used in this change to limit our locks to
Linux and Mac to pre-emptively avoid otherwise spurious lock errors.

ALl Python lock files are re-generated with the new `--target-system`
defaulting in place in order to drop a bunch of Windows-only
dependencies and artifacts.

See the changelog here:
  https://github.com/pantsbuild/pex/releases/tag/v2.1.95

[ci skip-rust]
[ci skip-build-wheels]