Make the "Merge OK" CI check work properly (#16095)
The previous attempt was broken because GHA treats a skipped job as if it passed.

Now we use the two "Set Merge OK" paths to set an output that another job checks for,
and that job is the one we actually have branch protection on. Only if one of the two
"Set Merge OK" jobs actually runs will that output be set.