Optimize `get_expanded_specs` for `StreamingWorkunit` plugins (#16106)
`StreamingWorkunit` plugins can use `get_expanded_specs` to capture the addresses associated with the input CLI specs for a run. But in cases with large numbers of input roots, this results in engine requests proportional to the number of targets in the repository.

This change adjusts the algorithm for building the mapping from unexpanded addresses to expanded targets, and cleans up the `Scheduler.execute` API to allow for making requests with different product types in parallel.

[ci skip-rust]
[ci skip-build-wheels]