Initial cleanup and simplification of DynamicRemoteOptions.from_options
Note that this removes the call to `bootstrap_options.remote_auth_plugin.strip()` which means the config option needs to be clean w/o any spaces since we no will no longer clean it up.

Context for this change:
I am refactoring this logic in order to add support for specifying the remote_auth_plugin via a backend/plugin entrypoint instead of an option which will simplify the config.