Use filelists to make archive creation tasks more reliable (#16183)
Previously, creating an archive with a large number of input files would cause the binary to fail due to having a too-long command line. 

This makes use of `tar`'s `--files-from` option to consume the input filenames from a file, and replaces our direct `zip` invocation with a bash script that provides the input filenames by stdin. 

There is no practical limit to the length of the file list that can be provided in this way.

Closes #16091
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]