Apply parameter filling for `parametrize`d dependencies to excludes, and fix accommodation for file-addressed atom targets (#16249)
As described in #16241, parameter filling for `parametrize`d dependencies is currently not applied to excludes, which makes it impossible to use excludes with a `parametrize`d dependee.

Additionally, the accommodation from #14419 that allows an atom target to be addressed by a file address was not being applied to parametrization, which caused the failure described in #16230.

Fixes #16241 and fixes #16230.

[ci skip-rust]
[ci skip-build-wheels]