[internal] Shell completion support (#16200)
Adding a script to generate tab-completion for Bash (and sorta Zsh, but that should be another custom script).

The available goals and options are generated by ./pants help-all. As a result, the completion should align with the pants.toml backend packages, and should be re-generated/sourced when that is changed.

Currently there is no target selection, nor is there any automation regarding re-builds or automatic sourcing. This PR is strictly related to the explicit generation of completions on demand. Separate PRs will be opened for the missing functionality as the completion functionality stabilizes.

(Closes issue #11187)

[ci skip-rust]
[ci skip-build-wheels]